+++ apps/trackers/jira/save.py
@@ -44,15 +44,16 @@ class TrackerJiraSaver(JiraQuerier):
         """
         create a representation of tracker model in Jira
         """
-        query = TrackerJiraQueryBuilder(tracker).generate()
+        query = TrackerJiraQueryBuilder(tracker).query
         issue = self.jira_conn.create_issue(fields=query["fields"], prefetch=True)
-        tracker.external_system_id = issue["key"]
+        tracker.external_system_id = issue.key
         return tracker
 
     def update(self, tracker):
         """
         update an existing representation of tracker model in Jira
         """
-        query = TrackerJiraQueryBuilder(tracker).generate()
+        query = TrackerJiraQueryBuilder(tracker).query
         url = f"{self.jira_conn._get_url('issue')}/{query['key']}"
         self.jira_conn._session.put(url, json.dumps(query))
+        return tracker