+++ docs/CHANGELOG.md
@@ -5,10 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Implement writable tracker API (OSIDB-1180)
+
 ### Fixed
 - Fix incorrect type bool of is_up2date field in
   /collectors/api/v1/status endpoint
-
 
 ## [3.5.0] - 2023-10-09
 ### Added